backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15.2,<3.16
# using the binary version to run docker compose fast
# But it's better to compile from source for better performance
psycopg2-binary>=2.9.9,<2.10
drf-spectacular>=0.27.2,<0.28
Pillow>=10.4.0,<10.5
django-cors-headers>=4.4.0,<4.5
faker>=27.0.0,<27.1 # Only in DEV
uvicorn>=0.30.6,<0.31
adrf>=0.1.7,<0.2
drf-nested-routers>=0.94.1,<0.95